requirements.txt
# Core dependencies
langgraph>=0.2.0
streamlit>=1.38.0
polygon-api-client>=1.14.0
duckduckgo-search>=6.3.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=8.3.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0
